+++ src/schemathesis/converter.py
@@ -1,5 +1,7 @@
 from copy import deepcopy
-from typing import Any, Dict, List, Union, overload
+from typing import Any, Dict
+
+from .utils import traverse_schema
 
 
 def to_json_schema(schema: Dict[str, Any], nullable_name: str) -> Dict[str, Any]:
@@ -25,40 +27,10 @@ def to_json_schema(schema: Dict[str, Any], nullable_name: str) -> Dict[str, Any]
     return schema
 
 
-Schema = Union[Dict[str, Any], List, str, float, int]
-
-
-@overload
 def to_json_schema_recursive(schema: Dict[str, Any], nullable_name: str) -> Dict[str, Any]:
-    pass
-
-
-@overload
-def to_json_schema_recursive(schema: List, nullable_name: str) -> List:
-    pass
-
-
-@overload
-def to_json_schema_recursive(schema: str, nullable_name: str) -> str:
-    pass
-
-
-@overload
-def to_json_schema_recursive(schema: float, nullable_name: str) -> float:
-    pass
-
-
-def to_json_schema_recursive(schema: Schema, nullable_name: str) -> Schema:
     """Apply ``to_json_schema`` recursively.
 
     This version is needed for cases where the input schema was not resolved and ``to_json_schema`` wasn't applied
     recursively.
     """
-    if isinstance(schema, dict):
-        schema = to_json_schema(schema, nullable_name)
-        for key, sub_item in schema.items():
-            schema[key] = to_json_schema_recursive(sub_item, nullable_name)
-    elif isinstance(schema, list):
-        for idx, sub_item in enumerate(schema):
-            schema[idx] = to_json_schema_recursive(sub_item, nullable_name)
-    return schema
+    return traverse_schema(schema, to_json_schema, nullable_name)