+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`1.4.0`_ - 2020-05-03
+---------------------
+
 Added
 ~~~~~
 
@@ -929,7 +932,8 @@ Fixed
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v1.3.4...HEAD
+.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v1.4.0...HEAD
+.. _1.4.0: https://github.com/kiwicom/schemathesis/compare/v1.3.4...v1.4.0
 .. _1.3.4: https://github.com/kiwicom/schemathesis/compare/v1.3.3...v1.3.4
 .. _1.3.3: https://github.com/kiwicom/schemathesis/compare/v1.3.2...v1.3.3
 .. _1.3.2: https://github.com/kiwicom/schemathesis/compare/v1.3.1...v1.3.2